// packages/deepspace/src/storage/types.ts
export interface RecordData<T = unknown> {
  recordId: string
  data: T
  createdBy: string
  createdAt: string
  updatedAt: string
}
// ...
export interface RoomUser {
  id: string
  email: string
  name: string
  imageUrl?: string
  role: string
  createdAt: string
  lastSeenAt: string
}
// ...
export type ConnectionStatus = 'connecting' | 'connected' | 'disconnected'